+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,13 +1,13 @@
 package ru.yandex.practicum.filmorate.controller;
 
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
-import jakarta.validation.Valid;
 import java.util.Set;
 
 @Slf4j
@@ -24,49 +24,62 @@ public class UserController {
 
     @PostMapping
     public User create(@Valid @RequestBody User user) {
-        log.info("Создан пользователь: {}", user.getLogin());
+        if (user.getName() == null || user.getName().trim().isEmpty()) {
+            user.setName(user.getLogin());
+        }
         return userService.create(user);
     }
 
     @PutMapping
     public User update(@Valid @RequestBody User user) {
-        log.info("Обновлён пользователь: {}", user.getLogin());
+        if (user.getName() == null || user.getName().trim().isEmpty()) {
+            user.setName(user.getLogin());
+        }
         return userService.update(user);
     }
 
     @GetMapping
     public Set<User> getAll() {
-        log.info("Запрос на получение всех пользователей.");
         return userService.getAll();
     }
 
     @GetMapping("/{id}")
     public User getById(@PathVariable int id) {
-        log.info("Запрос пользователя с id={}", id);
+        if (id <= 0) {
+            throw new ValidationException("ID пользователя должен быть положительным.");
+        }
         return userService.getById(id);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void addFriend(@PathVariable int id, @PathVariable int friendId) {
+        if (id <= 0 || friendId <= 0) {
+            throw new ValidationException("ID пользователя должен быть положительным.");
+        }
         userService.addFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void removeFriend(@PathVariable int id, @PathVariable int friendId) {
+        if (id <= 0 || friendId <= 0) {
+            throw new ValidationException("ID пользователя должен быть положительным.");
+        }
         userService.removeFriend(id, friendId);
     }
 
     @GetMapping("/{id}/friends")
     public Set<User> getFriends(@PathVariable int id) {
-        log.info("Запрос друзей пользователя с id={}", id);
+        if (id <= 0) {
+            throw new ValidationException("ID пользователя должен быть положительным.");
+        }
         return userService.getFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
     public Set<User> getCommonFriends(@PathVariable int id, @PathVariable int otherId) {
-        log.info("Запрос общих друзей между {} и {}", id, otherId);
+        if (id <= 0 || otherId <= 0) {
+            throw new ValidationException("ID пользователя должен быть положительным.");
+        }
         return userService.getCommonFriends(id, otherId);
     }
 }
